double vs int in an eternal fight

+++ cnF2freq.cpp
@@ -2694,7 +2694,7 @@ struct genotypereporter : valuereporter<4>
 
 struct statereporter : valuereporter<NUMTYPES>
 {
-	void addval(int q, int mapval, int g, int flag2, int val)
+	void addval(int q, int mapval, int g, int flag2, double val)
 	{
 		probs[g] += val;
 	}
